# openapi.yaml
openapi: 3.0.3
info:
  title: J1Hub API
  version: 1.0.0
  description: |
    Stable API contract for J1Hub endpoints.

    ## Versioning plan
    - Canonical versioned base path: `/api/v1`.
    - New clients use `/api/v1/*`; unversioned routes remain available while `API_ENABLE_LEGACY_ROUTES=true`.
    - New clients should migrate to `/api/v1/*` immediately.

    ## Transition strategy
    1. **Now**: support both `/api/v1/*` and legacy endpoints.
    2. **Deprecation window**: communicate deprecation for legacy routes via release notes and docs.
    3. **Removal**: disable legacy aliases by configuration after migration completion.
servers:
  - url: /
    description: Legacy (transition period)
  - url: /api/v1
    description: Canonical versioned API
security:
  - bearerAuth: []
paths:
...
  /listings:
    get:
      summary: Search listings
      operationId: searchListings
      security:
        - {}
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [job, housing, ride, gig]
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: active
          schema:
            oneOf:
              - type: boolean
              - type: string
      responses:
        '200':
          description: Listings search results
          content:
            application/json:
              schema:
                type: object
                required: [results, count]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listing'
                  count:
                    type: integer
                  pagination:
                    type: object
        '400':
          $ref: '#/components/responses/ErrorResponse'
    post:
      summary: Create listing
      operationId: createListing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        '201':
          description: Created listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '402':
          description: Billing required
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/mine:
    get:
      summary: List the authenticated employer's listings
      operationId: listMyListings
      responses:
        '200':
          description: Owned listings
          content:
            application/json:
              schema:
                type: object
                required: [results, count]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listing'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/{listing_id}:
    get:
      summary: Get listing
      operationId: getListing
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      summary: Update listing
      operationId: updateListing
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdateRequest'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/{listing_id}/apply:
    post:
      summary: Apply to listing
      operationId: applyToListing
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/{listing_id}/applications:
    get:
      summary: List applications for an employer-owned listing
      operationId: listListingApplications
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Applications for the listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/applications/mine:
    get:
      summary: List the authenticated worker's applications
      operationId: listMyApplications
      responses:
        '200':
          description: Applications submitted by the worker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/{listing_id}/applications/{application_id}:
    patch:
      summary: Update application status
      operationId: updateApplicationStatus
      parameters:
        - $ref: '#/components/parameters/ListingId'
        - in: path
          name: application_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [new, reviewing, accepted, rejected]
      responses:
        '200':
          description: Updated application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/favorites:
    get:
      summary: List the authenticated worker's saved listings
      operationId: listFavoriteListings
      responses:
        '200':
          description: Saved listings
          content:
            application/json:
              schema:
                type: object
                required: [results, count]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /listings/{listing_id}/favorite:
    post:
      summary: Save a listing
      operationId: saveListing
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Existing saved listing
          content:
            application/json:
              schema:
                type: object
        '201':
          description: Saved listing created
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Remove a saved listing
      operationId: removeSavedListing
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Saved listing removed
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [deleted]
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListingId:
      in: path
      name: listing_id
      required: true
      schema:
        type: integer
...
  responses:
    ErrorResponse:
      description: JSON error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error, detail, request_id]
      properties:
        error:
          type: string
        detail:
          type: string
        request_id:
          type: string
...
    Listing:
      type: object
      required:
        [id, category, title, description, is_public, is_active, created_by]
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [job, housing, ride, gig]
        title:
          type: string
        description:
          type: string
        company_name:
          type: string
          nullable: true
        contact_method:
          type: string
          enum: [phone, email, in_app]
          nullable: true
        contact_value:
          type: string
          nullable: true
        location_city:
          type: string
          nullable: true
        pay_rate:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        shift:
          type: string
          nullable: true
        is_public:
          type: boolean
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
    ListingCreateRequest:
      type: object
      required: [category, title, description, contact_method, contact_value]
      properties:
        category:
          type: string
          enum: [job, housing, ride, gig]
        title:
          type: string
        description:
          type: string
        company_name:
          type: string
        contact_method:
          type: string
          enum: [phone, email, in_app]
        contact_value:
          type: string
        location_city:
          type: string
        pay_rate:
          oneOf:
            - type: number
            - type: string
        currency:
          type: string
        shift:
          type: string
        is_public:
          type: boolean
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
    ListingUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        category:
          type: string
          enum: [job, housing, ride, gig]
        title:
          type: string
        description:
          type: string
        company_name:
          type: string
        contact_method:
          type: string
          enum: [phone, email, in_app]
        contact_value:
          type: string
        location_city:
          type: string
        pay_rate:
          oneOf:
            - type: number
            - type: string
            - type: 'null'
        currency:
          type: string
        shift:
          type: string
        is_public:
          type: boolean
        is_active:
          type: boolean
        expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    Application:
      type: object
      required: [id, user_id, listing_id, message, status, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        listing_id:
          type: integer
        message:
          type: string
        status:
          type: string
          enum: [new, reviewing, accepted, rejected]
        created_at:
          type: string
          format: date-time
          nullable: true
...
    ApplicationList:
      type: object
      required: [results, count]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        count:
          type: integer
...